(MTR): явно указал типа аргументов у функции pow(), чтобы не ругался компилятор в p6.

+++ extensions/include/MTR.h
@@ -242,7 +242,7 @@ namespace MTR
 			{
 				raw.v[0] = v1;
 				raw.v[1] = v2;
-				val = raw.u2.val * pow(10,raw.u2.exp);
+				val = raw.u2.val * pow( (long)10, (long)raw.u2.exp );
 			}
 			
 			T5( long v )
@@ -257,7 +257,7 @@ namespace MTR
 					// При получении данных от MTR слова необходимо перевернуть
 					raw.v[0] = data[1];
 					raw.v[1] = data[0];
-					val = raw.u2.val * pow(10,raw.u2.exp);
+					val = raw.u2.val * pow( (long)10, (long)raw.u2.exp );
 				}
 			}
 
@@ -295,7 +295,7 @@ namespace MTR
 			{
 				raw.v[0] = v1;
 				raw.v[1] = v2;
-				val = raw.u2.val * pow(10,raw.u2.exp);
+				val = raw.u2.val * pow( (long)10, (long)raw.u2.exp );
 			}
 
 			T6( long v )
@@ -310,7 +310,7 @@ namespace MTR
 					// При получении данных от MTR слова необходимо перевернуть
 					raw.v[0] = data[1];
 					raw.v[1] = data[0];
-					val = raw.u2.val * pow(10,raw.u2.exp);
+					val = raw.u2.val * pow( (long)10, (long)raw.u2.exp );
 				}
 			}
 
@@ -349,7 +349,7 @@ namespace MTR
 			{
 				raw.v[0] = v1;
 				raw.v[1] = v2;
-				val = raw.u2.val * pow(10,-4);
+				val = raw.u2.val * pow( (long)10, (long)-4 );
 			}
 			T7( const long v )
 			{
@@ -363,7 +363,7 @@ namespace MTR
 					// При получении данных от MTR слова необходимо перевернуть
 					raw.v[0] = data[1];
 					raw.v[1] = data[0];
-					val = raw.u2.val * pow(10,-4);
+					val = raw.u2.val * pow( (long)10, (long)-4 );
 				}
 			}
